% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VoterModel.R
\name{CVM_extr}
\alias{CVM_extr}
\title{Confidence voter model - extremal}
\usage{
CVM_extr(row, col, s)
}
\arguments{
\item{row}{number of rows}

\item{col}{number of columns}

\item{s}{number of observations}
}
\value{
Time to consensus of each observation
}
\description{
Assumes boundaries
}
