% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VoterModel.R
\name{CVM_extr_V2}
\alias{CVM_extr_V2}
\title{Confidence voter model - extremal (Version 2)}
\usage{
CVM_extr_V2(row, col, s, change = 1)
}
\arguments{
\item{row}{Number of rows}

\item{col}{Number of columns}

\item{s}{Number of observations}

\item{change}{Probability that an individual will be convinced by a confident neighbor (constant for each individual). Default is 1}
}
\value{
Time to consensus of each observation
}
\description{
Runs similarly to CVM_extr, but state/confidence only change if neighbor is confident. We can also adjust the probability that an individual is convinced by a neighbor (conditioned on that neighbor being confident). Assumes boundaries
}
