% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VoterModel.R
\name{CVM_func_V2}
\alias{CVM_func_V2}
\title{CVM_func_V2}
\usage{
CVM_func_V2(fixed.row, col.range, change = 1)
}
\arguments{
\item{fixed.row}{Number of rows that will remain fixed through each simulation}

\item{col.range}{Range of columns. Each will be used twice: once in marginal, and once in extremal case.}

\item{change}{Probability that an individual will be convinced by a confident neighbor (constant for each individual). Default is 1}

\item{returns}{Average of, variance of, second moment of consensus of each dimension, for both marginal and extremal cases (6 vectors total)}
}
\description{
Models the time to consensus of the confidence voter models (Version 2) as a function over a number of column
}
