% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VoterModel.R
\name{CVM_marg_V3}
\alias{CVM_marg_V3}
\title{Confidence voter model - marginal (Version 3)}
\usage{
CVM_marg_V3(row, col, s, pOne = 0.5)
}
\arguments{
\item{row}{Number of rows}

\item{col}{Number of columns}

\item{s}{Number of observations}

\item{pOne}{Probability that a spot in the matrix is initialized with '1'. Default is 0.5}
}
\value{
Time to consensus of each observation
}
\description{
Each individual will have their own probability of being influenced by a confident voter. Assumes boundaries, and only classic movements
}
