% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VoterModel.R
\name{myModified_percent}
\alias{myModified_percent}
\title{My modified voter model, percents}
\usage{
myModified_percent(row, col)
}
\arguments{
\item{row}{Number of rows}

\item{col}{Number of columns}

\item{pOne}{Probability that a spot in the matrix is initialized with '1'. Default is 0.5}
}
\value{
: Percent of population labeled "1" throughout the observation
}
\description{
Used in VM theory.R file. Here, the probability of being convinced is dependent on the proportion of agreeing voters. That is, if we have 3 individuals labeled "0" and 7 voters labeled "1", a "0" voter has a probability of 3/10 of being convinced, and a "1" voter has a probability of 7/10 of being convinced. We only run 1 observation. Assumes boundaries
}
