% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VoterModel.R
\name{myModified_time}
\alias{myModified_time}
\title{My modified voter model, times}
\usage{
myModified_time(row, col, s, pOne = 0.5)
}
\arguments{
\item{row}{Number of rows}

\item{col}{Number of columns}

\item{s}{Number of observations}

\item{pOne}{Probability that a spot in the matrix is initialized with '1'. Default is 0.5}
}
\value{
Time to consensus of each observation
}
\description{
used in VM theory.R file. Here, the probability of being convinced is dependent on the proportion of agreeing voters. That is, if we have 3 individuals labeled "0" and 7 voters labeled "1", a "0" voter has a probability of 3/10 of being convinced, and a "1" voter has a probability of 7/10 of being convinced. Assumes boundaries
}
