% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VoterModel.R
\name{neighbor}
\alias{neighbor}
\title{neighbor}
\usage{
neighbor(nbhd, i, j, boundaries = TRUE)
}
\arguments{
\item{nbhd}{Matrix representing the different individuals}

\item{i}{Row number of individual}

\item{j}{Column number of individual}

\item{boundaries}{If TRUE, we assume boundaries, otherwise edges and corners are linked to the other side of the matrix. Default is set to TRUE.}
}
\value{
Row and column number of a randomly selected neighbor
}
\description{
Uniformly selects neighbor of an individual using classic movements
}
