% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VoterModel.R
\name{VM_complete}
\alias{VM_complete}
\title{Classic model on a complete graph}
\usage{
VM_complete(row, col, s, pOne = 0.5)
}
\arguments{
\item{row}{Number of rows}

\item{col}{Number of columns}

\item{s}{Number of observations}

\item{pOne}{Probability that a spot in the matrix is initialized with '1'. Default is 0.5}
}
\value{
Time to consensus of each observation
}
\description{
Looks at the classic voter model, but on a complete graph, rather than a square lattice graph.
}
