% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VoterModel.R
\name{VM_complete_func}
\alias{VM_complete_func}
\title{VM_complete_func}
\usage{
VM_complete_func(row, col, s)
}
\arguments{
\item{row}{Number of rows}

\item{col}{Number of columns}

\item{s}{Number of observations}
}
\value{
Average time to consensus of each initial density
}
\description{
Models the time to consensus of VM_complete as a function over the initial density of agreeing voters.
}
