% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VoterModel.R
\name{VM_func}
\alias{VM_func}
\title{VM_func}
\usage{
VM_func(fixed.row, col.range)
}
\arguments{
\item{fixed.row}{Number of rows that will remain fixed through each simulation}

\item{col.range}{Range of columns}
}
\value{
Average of, variance of, second moment of consensus of each dimension (3 vectors total)
}
\description{
Models the time to consensus of VM as a function over a number of columns
}
